requirements.txt
numpy==2.3.1
pillow==11.3.0
PyOpenGL==3.1.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
trimesh==4.7.1
